--- components/Behaviors/samples/BehaviorsXbindBackedStyleCustomSample.xaml
@@ -1,28 +1,0 @@
-<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
-<Page
-    x:Class="BehaviorsExperiment.Samples.BehaviorsXbindBackedStyleCustomSample"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Page.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <!--  Import custom styles  -->
-                <controls:BehaviorsStyle_xBind />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Page.Resources>
-
-    <StackPanel Spacing="16">
-        <controls:Behaviors_xBind
-            FontFamily="{x:Bind TextFontFamily, Mode=OneWay}"
-            FontSize="{x:Bind TextSize, Mode=OneWay}"
-            Foreground="{x:Bind TextForeground, Mode=OneWay}"
-            Style="{StaticResource CustomxBindBehaviorsStyle}"
-            Visibility="{x:Bind IsTextVisible, Mode=OneWay}" />
-    </StackPanel>
-</Page>
