+++ components/Behaviors/samples/LoadMoreItemBehaviorSample.xaml
@@ -1,10 +1,9 @@
 <!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
 <Page
-    x:Class="BehaviorsExperiment.Samples.NeedMoreItemTriggerBehaviorSample"
+    x:Class="BehaviorsExperiment.Samples.LoadMoreItemBehaviorSample"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:CommunityToolkit.WinUI.Controls"
-    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:local="using:BehaviorsExperiment.Samples"
@@ -27,22 +26,19 @@
                     MinRowSpacing="5" />
             </ItemsView.Layout>
             <interactivity:Interaction.Behaviors>
-                <controls:NeedMoreItemTriggerBehavior LoadingOffset="{x:Bind LoadingOffset, Mode=OneWay}">
-                    <core:CallMethodAction MethodName="IncrementCount" TargetObject="{x:Bind}" />
-                </controls:NeedMoreItemTriggerBehavior>
-
+                <controls:LoadMoreItemBehavior LoadingOffset="{x:Bind LoadingOffset, Mode=OneWay}" />
             </interactivity:Interaction.Behaviors>
             <ItemsView.ItemTemplate>
                 <DataTemplate x:DataType="x:Int32">
                     <ItemContainer>
-                        <Rectangle Width="{x:Bind}" Fill="{x:Bind local:NeedMoreItemTriggerBehaviorSample.GetColor()}" />
+                        <Rectangle Width="{x:Bind}" Fill="{x:Bind local:LoadMoreItemBehaviorSample.GetColor()}" />
                     </ItemContainer>
                 </DataTemplate>
             </ItemsView.ItemTemplate>
         </ItemsView>
         <TextBlock Grid.Column="1" TextWrapping="Wrap">
-            <Run Text="CallMethodAction triggered " />
-            <Run x:Name="MyRun" Text="0" />
+            <Run Text="Load more triggered " />
+            <Run Text="{x:Bind ViewModels.Num, Mode=OneWay}" />
             <Run Text=" times" />
         </TextBlock>
     </Grid>
