+++ components/Behaviors/samples/NeedMoreItemTriggerBehaviorSample.xaml.cs
@@ -2,20 +2,19 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
+#if WINAPPSDK
+
 using CommunityToolkit.WinUI.Controls;
+using Microsoft.UI;
 
 namespace BehaviorsExperiment.Samples;
 
-/// <summary>
-/// An example sample page of a custom control inheriting from Panel.
-/// </summary>
-[ToolkitSampleNumericOption("LoadingOffset", 100d, 500d, 50d, Title = "LoadingOffset")]
+[ToolkitSampleNumericOption("LoadingOffset", 100d, 50d, 500d, 50d, Title = "LoadingOffset")]
 
-[ToolkitSample(id: nameof(NeedMoreItemTriggerBehaviorSample), "Custom control", description: $"A sample for showing how to create and use a {nameof(Behaviors)} custom control.")]
+[ToolkitSample(id: nameof(NeedMoreItemTriggerBehaviorSample), "NeedMoreItemTriggerBehavior", description: $"A sample for showing how to create and use a NeedMoreItemTriggerBehavior.")]
 public sealed partial class NeedMoreItemTriggerBehaviorSample : Page
 {
-    private static readonly Random _random = new Random();
+    private static readonly Random _random = new();
 
     public NeedMoreItemTriggerBehaviorSample()
     {
@@ -36,46 +35,15 @@ public sealed partial class NeedMoreItemTriggerBehaviorSample : Page
         return new(color);
     }
 
-    public MyCollection ViewModels { get; } = [.. Enumerable.Repeat(0, 50).Select(_ => _random.Next(5, 10) * 500)];
-}
-
-public partial class MyCollection : ObservableCollection<int>, ISupportIncrementalLoading
-{
-    private static readonly Random _random = new Random();
-
-    public int Num
-    {
-        get => _num;
-        set
-        {
-            if (_num == value)
-            {
-                return;
-            }
-
-            _num = value;
-            OnPropertyChanged(new PropertyChangedEventArgs(nameof(Num)));
-        }
-    }
+    public IEnumerable<int> ViewModels { get; } = Enumerable.Repeat(0, 50).Select(_ => _random.Next(5, 10) * 500);
 
     private int _num;
 
-    public bool HasMoreItems => true;
-
-    public IAsyncOperation<LoadMoreItemsResult> LoadMoreItemsAsync(uint count)
+    public void IncrementCount()
     {
-        return LoadMoreItemsTaskAsync(count).AsAsyncOperation();
-    }
-
-    public async Task<LoadMoreItemsResult> LoadMoreItemsTaskAsync(uint count)
-    {
-        await Task.Delay(1000);
-        foreach (var i in Enumerable.Repeat(0, (int)count).Select(_ => _random.Next(5, 10) * 500))
-        {
-            Add(i);
-        }
-
-        Num++;
-        return new LoadMoreItemsResult { Count = count };
+        _num++;
+        MyRun.Text = _num.ToString();
     }
 }
+
+#endif
