--- components/Behaviors/src/BehaviorsStyle_ClassicBinding.xaml
@@ -1,62 +1,0 @@
-﻿<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
-<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:controls="using:CommunityToolkit.WinUI.Controls">
-
-    <ResourceDictionary.ThemeDictionaries>
-        <ResourceDictionary x:Key="Light">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="#000000" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="Dark">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="{StaticResource SystemBaseMediumColor}" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="HighContrast">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="{ThemeResource SystemColorButtonFaceColor}" />
-        </ResourceDictionary>
-    </ResourceDictionary.ThemeDictionaries>
-
-    <Thickness x:Key="DefaultBehaviorsItemPadding">4,4,4,4</Thickness>
-
-    <!--  Implicitly applied default style  -->
-    <Style BasedOn="{StaticResource DefaultClassicBindingBehaviorsStyle}"
-           TargetType="controls:Behaviors_ClassicBinding" />
-
-    <Style x:Key="DefaultClassicBindingBehaviorsStyle"
-           TargetType="controls:Behaviors_ClassicBinding">
-        <Style.Setters>
-            <Setter Property="MyProperty" Value="This property has been overriden by the style template" />
-            <Setter Property="ItemPadding" Value="{StaticResource DefaultBehaviorsItemPadding}" />
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="controls:Behaviors_ClassicBinding">
-                        <Grid Padding="{TemplateBinding ItemPadding}">
-                            <StackPanel Spacing="20">
-                                <TextBlock x:Name="PART_HelloWorld"
-                                           Foreground="{ThemeResource MyBrush}"
-                                           Text="Hello world!" />
-
-                                <TextBlock Text="{TemplateBinding MyProperty}" />
-                                <TextBlock Text="{Binding MyProperty, RelativeSource={RelativeSource Mode=TemplatedParent}}"
-                                           Visibility="Collapsed" />
-                            </StackPanel>
-                        </Grid>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style.Setters>
-    </Style>
-
-    <!--  Minimal, custom style  -->
-    <Style x:Key="CustomClassicBindingBehaviorsStyle"
-           BasedOn="{StaticResource DefaultClassicBindingBehaviorsStyle}"
-           TargetType="controls:Behaviors_ClassicBinding">
-        <Style.Setters>
-            <Setter Property="MyProperty" Value="This property has been overriden by a custom style!" />
-            <Setter Property="ItemPadding" Value="25" />
-        </Style.Setters>
-    </Style>
-
-</ResourceDictionary>
