--- components/Behaviors/src/BehaviorsStyle_xBind.xaml
@@ -1,69 +1,0 @@
-﻿<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
-<ResourceDictionary x:Class="CommunityToolkit.WinUI.Controls.BehaviorsStyle_xBind"
-                    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:controls="using:CommunityToolkit.WinUI.Controls">
-
-    <ResourceDictionary.ThemeDictionaries>
-        <ResourceDictionary x:Key="Light">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="#000000" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="Dark">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="{StaticResource SystemBaseMediumColor}" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="HighContrast">
-            <SolidColorBrush x:Key="MyBrush"
-                             Color="{ThemeResource SystemColorButtonFaceColor}" />
-        </ResourceDictionary>
-    </ResourceDictionary.ThemeDictionaries>
-
-    <Thickness x:Key="DefaultBehaviorsItemPadding">4,4,4,4</Thickness>
-
-    <!--  Implicitly applied default style  -->
-    <Style BasedOn="{StaticResource DefaultxBindBehaviorsStyle}"
-           TargetType="controls:Behaviors_xBind" />
-
-    <Style x:Key="DefaultxBindBehaviorsStyle"
-           TargetType="controls:Behaviors_xBind">
-        <Style.Setters>
-            <Setter Property="MyProperty" Value="This property has been overriden by the style template" />
-            <Setter Property="ItemPadding" Value="{StaticResource DefaultBehaviorsItemPadding}" />
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="controls:Behaviors_xBind">
-                        <ContentControl x:Name="ContentContainer"
-                                        HorizontalContentAlignment="Stretch"
-                                        VerticalContentAlignment="Stretch">
-                            <ContentControl.ContentTemplate>
-                                <DataTemplate x:DataType="controls:Behaviors_xBind">
-                                    <Grid Padding="{x:Bind ItemPadding}"
-                                          PointerEntered="{x:Bind Element_PointerEntered}">
-                                        <StackPanel Spacing="20">
-                                            <TextBlock Foreground="{ThemeResource MyBrush}"
-                                                       Text="Hello world!" />
-
-                                            <TextBlock Text="{x:Bind MyProperty, Mode=OneWay}" />
-                                        </StackPanel>
-                                    </Grid>
-                                </DataTemplate>
-                            </ContentControl.ContentTemplate>
-                        </ContentControl>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style.Setters>
-    </Style>
-
-    <!--  Minimal, custom style  -->
-    <Style x:Key="CustomxBindBehaviorsStyle"
-           BasedOn="{StaticResource DefaultxBindBehaviorsStyle}"
-           TargetType="controls:Behaviors_xBind">
-        <Style.Setters>
-            <Setter Property="MyProperty" Value="This property has been overriden by a custom style!" />
-            <Setter Property="ItemPadding" Value="25" />
-        </Style.Setters>
-    </Style>
-
-</ResourceDictionary>
