--- components/Behaviors/src/BehaviorsStyle_xBind.xaml.cs
@@ -1,20 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace CommunityToolkit.WinUI.Controls;
-
-/// <summary>
-/// Backing code for this resource dictionary.
-/// </summary>
-public sealed partial class BehaviorsStyle_xBind : ResourceDictionary
-{
-    // NOTICE
-    // This file only exists to enable x:Bind in the resource dictionary.
-    // Do not add code here.
-    // Instead, add code-behind to your templated control.
-    public BehaviorsStyle_xBind()
-    {
-        this.InitializeComponent();
-    }
-}
